requirements.txt
faiss-cpu==1.11.0.post1
openai==1.97.1
tiktoken>=0.4.0
fastapi==0.116.1
uvicorn==0.35.0
gspread==6.2.1
google-api-python-client==2.176.0
oauth2client==4.1.3
pydantic>=2.11.7
pdfplumber==0.11.7
google-auth
numpy>=1.21.0
langdetect==1.0.9
azure-ai-translation-text==1.0.1
python-dotenv==1.0.1
azure-core
requests
